Update Bus to Udayland.cpp

+++ 800/Bus to Udayland.cpp
@@ -3,8 +3,8 @@
 using namespace std;
 
 int main() 
-{
-	int n;
+{ 
+        ll n;
 	cin>>n;
 	char mat[n][5];
 	for(int i=0;i<n;i++)
